Refactored JoinDutyFinder, moved ExtendedContentsFinder
